requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
anthropic==0.39.0
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.29.0
plotly==5.18.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.11
PyPDF2==3.0.1
httpx==0.25.2
python-multipart==0.0.6
fpdf2==2.7.6
shap==0.44.0
numpy==1.26.2
aiofiles==23.2.1
python-dotenv==1.0.0
